Issue #10 Add play and download button when hovering on movie image
I have created play and download button which will appear on hover.

Also made the code much more easier to read by adding comments, indents and spaces cause it was becoming difficult to work on this kind of large project 😅 Hope It is fine.

Please let me know if you need any changes to be done